@@ -1,6 +1,6 @@
 // React imports
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 
 // App css imports
@@ -14,13 +14,14 @@ import './index.scss';
 import App from './App';
 import { AppProvider } from './state-management/stores/Store';
 
-ReactDOM.render(
+const container = document.getElementById('root') as HTMLElement;
+const rootContainer = ReactDOM.createRoot(container);
+rootContainer.render(
     <React.StrictMode>
         <AppProvider>
             <BrowserRouter basename="/ui">
                 <App />
             </BrowserRouter>
         </AppProvider>
-    </React.StrictMode>,
-    document.getElementById('root')
+    </React.StrictMode>
 );
